+++ src/main/java/io/github/depromeet/knockknockbackend/domain/group/service/InviteService.java
@@ -1,6 +1,8 @@
 package io.github.depromeet.knockknockbackend.domain.group.service;
 
+import io.github.depromeet.knockknockbackend.domain.group.domain.Group;
 import io.github.depromeet.knockknockbackend.domain.group.domain.repository.InviteRepository;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -9,4 +11,7 @@ import org.springframework.stereotype.Service;
 public class InviteService {
 
     private final InviteRepository inviteRepository;
+
+    public void makeInvites(Group group, List<Long> sendInviteUserIds, Long reqUserId) {
+    }
 }
